# openapi.yaml
openapi: 3.1.0
info:
  title: Job Search Dashboard API
  version: 0.1.0
  description: REST API for the Job Search + Resume Management Dashboard
servers:
  - url: http://localhost:3000
    description: Local development
paths:
...
  /api/tasks:
    get:
      operationId: listTasks
      summary: List tasks
      tags: [Tasks]
      parameters:
        - name: done
          in: query
          schema:
            type: string
            enum: ['true', 'false']
      responses:
        '200':
          description: List of tasks
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Task'
    post:
      operationId: createTask
      summary: Create task
      tags: [Tasks]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTask'
      responses:
        '201':
          description: Created task
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Task'
        '400':
          $ref: '#/components/responses/ValidationError'
  /api/tasks/{id}:
    parameters:
      - $ref: '#/components/parameters/Id'
    put:
      operationId: updateTask
      summary: Update task
      tags: [Tasks]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTask'
      responses:
        '200':
          description: Updated task
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Task'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      operationId: deleteTask
      summary: Delete task
      tags: [Tasks]
      responses:
        '200':
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      id:
                        type: string
        '404':
          $ref: '#/components/responses/NotFound'
  /api/settings:
    get:
      operationId: listSettings
      summary: List all settings
      tags: [Settings]
      responses:
        '200':
          description: List of settings
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Setting'
    post:
      operationId: upsertSetting
      summary: Upsert setting by key
      tags: [Settings]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpsertSetting'
      responses:
        '201':
          description: Upserted setting
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Setting'
        '400':
          $ref: '#/components/responses/ValidationError'
...
components:
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: string
...
  responses:
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
              details:
                type: object
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
  schemas:
...
    Task:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        title:
          type: string
        done:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateTask:
      type: object
      required: [title]
      properties:
        title:
          type: string
          minLength: 1
        done:
          type: boolean
          default: false
    UpdateTask:
      type: object
      properties:
        title:
          type: string
        done:
          type: boolean
    Setting:
      type: object
      properties:
        id:
          type: string
        key:
          type: string
        value:
          type: string
        updatedAt:
          type: string
          format: date-time
    UpsertSetting:
      type: object
      required: [key, value]
      properties:
        key:
          type: string
          minLength: 1
        value:
          type: string
...